+++ collabora-platform-extension/src/main/java/fr/jeci/collabora/wopi/WopiPutFileWebScript.java
@@ -128,7 +128,7 @@ public class WopiPutFileWebScript extends AbstractWopiWebScript {
 	 */
 	private void checkWopiTimestamp(final WebScriptRequest req, final WebScriptResponse res, final NodeRef nodeRef)
 			throws IOException {
-		final Version currentVersion = getOrRepairCurrentVersion(nodeRef);
+		final Version currentVersion = getCurrentVersion(nodeRef);
 
 		if (currentVersion != null) {
 			// Check if X-LOOL-WOPI-Timestamp